ai-model/requirements.txt
# AI Model Dependencies
# Python 3.8+ required

# Core ML Libraries
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0

# Image Processing
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-image==0.21.0

# Data Handling
pandas==2.0.3
h5py==3.9.0

# API Server
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
waitress==2.1.2

# Utilities
tqdm==4.65.0
matplotlib==3.7.2
python-dotenv==1.0.0

# Model utilities
keras==2.13.1
